requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
webcolors